"Delete React Project
the react Project previously started is having troubles to boot up. Since the project had no progress
a new project will be started instead, it's expected to have this project running soon.

Move Flask Back End

Flask application was moved to a new folder, since it's no longer needed it will be move to a Legacy with
the rest of no longer needed files
"